third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/send-redirect-infinite-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: send() - Redirects (infinite loop) assert_equals: expected true but got false
Harness: the test ran to completion.
